requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
astunparse==1.6.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.119.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
httptools==0.7.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
keras==3.11.3
kiwisolver==1.4.9
langdetect==1.0.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mplcursors==0.7
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
onnx==1.19.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
RapidFuzz==3.14.1
requests==2.32.5
rich==14.2.0
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
validators==0.35.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
